Refactor code structure: update namespaces, improve error handling, and enhance user experience in views

+++ BlogAspNet.Web/Models/Services/CategoryService.cs
@@ -1,14 +1,18 @@
+using BlogAspNet.Web.Models.Entities;
 using BlogAspNet.Web.Models.Repositories;
+using BlogAspNet.Web.Models.Services.ViewModels;
 
 namespace BlogAspNet.Web.Models.Services;
 
 public class CategoryService : ICategoryService
 {
     private readonly ICategoryRepository _categoryRepository;
+    private readonly IBlogService _blogService;
 
-    public CategoryService(ICategoryRepository categoryRepository)
+    public CategoryService(ICategoryRepository categoryRepository, IBlogService blogService)
     {
         _categoryRepository = categoryRepository;
+        _blogService = blogService;
     }
 
     public async Task<bool> AddCategoryAsync(Category category)
@@ -23,7 +27,33 @@ public class CategoryService : ICategoryService
 
     public async Task<bool> DeleteCategoryAsync(int id)
     {
-        return await _categoryRepository.DeleteAsync(id);
+        if (id == 1)
+        {
+            return false;
+        }
+        var category = await _categoryRepository.GetByIdAsync(id);
+        if (category == null)
+        {
+            return false;
+        }
+
+        try
+        {
+            var blogs= await _blogService.GetBlogsByCategory(id);
+            foreach (var blog in blogs)
+            {
+                
+                var blogEditViewModel = await _blogService.GetBlogEditViewModel(blog.Id);
+                blogEditViewModel.CategoryId = 1; // Set to default category
+                await _blogService.UpdateBlog(blogEditViewModel);
+            }
+            await _categoryRepository.DeleteAsync(id);
+            return true;
+        }
+        catch (Exception)
+        {
+            return false;
+        }
     }
 
     public async Task<Category> GetCategoryByIdAsync(int id)
